+++ railties/lib/rails/generators/rails/poro_model/templates/model.rb
@@ -1,0 +1,4 @@
+<% module_namespacing do -%>
+class <%= class_name %>
+end
+<% end -%>