+++ railties/CHANGELOG.md
@@ -1,3 +1,19 @@
+*   Enable ModelGenerator `--no-orm` option to create PORO model
+
+    Creates a model as Plain Old Ruby Object:
+
+      $ bin/rails generate model address --no-orm
+
+    *Yoshiyuki Hirano*
+
+*   Add `Rails::Generators::PoroModelGenerator`.
+
+    Creates a model as Plain Old Ruby Object:
+
+      $ bin/rails generate poro_model address
+
+    *Yoshiyuki Hirano*
+
 *   Add `mini_magick` to default `Gemfile` as comment.
 
     *Yoshiyuki Hirano*